ONGOING ISSUE: Database Connection Problem
There is a database connection issue that I've been trying to resolve. A new database was created and connected to resolve a previous issue, and migrations files where deleted from apps, apart from __init__.py, but now the new database URL doesn't seem to connect to the project and I can't makemigrations and migrate.

This issue is under investigation.