FIX: Fix issue with order confirmation email
This commit deletes a piece of code that was causing issues with the order confirmation email. The deleted code was related to

@register.filter(name="none_to_empty")
def none_to_empty(value):
"""
Converts None or empty values to a default string.
This utility function is used to handle cases where a field's value
might be None or an empty string. It ensures that instead of displaying
'None' or leaving a field blank,
a default message ('Not provided') is returned.
Args:
value (str or None): The value to be checked.
It can be a string or None.
Returns:
str: The original value if it's not None or empty;
otherwise, the string 'Not provided'.
"""
return value if value else "Not provided"

Result: The issue with the Order confirmation email that wasn't working should be fixed now.